+++ code/src/sneer/bricks/snapps/games/go/impl/GoMainImpl.java
@@ -19,15 +19,15 @@ import sneer.bricks.snapps.contacts.actions.ContactAction;
 import sneer.bricks.snapps.contacts.actions.ContactActionManager;
 import sneer.bricks.snapps.contacts.gui.ContactsGui;
 import sneer.bricks.snapps.games.go.GoMain;
-import sneer.bricks.snapps.games.go.impl.gui.GoFrame;
+import sneer.bricks.snapps.games.go.impl.gui.GuiPlayer;
 import sneer.bricks.snapps.games.go.impl.logging.GoLogger;
 import sneer.bricks.snapps.games.go.impl.logic.GoBoard.StoneColor;
 import sneer.bricks.snapps.games.go.impl.logic.Move;
 import sneer.bricks.snapps.games.go.impl.network.GoInvitation;
 import sneer.bricks.snapps.games.go.impl.network.GoMessage;
 import sneer.bricks.snapps.games.go.impl.network.GoMove;
-import sneer.bricks.snapps.games.go.impl.network.Player;
-import sneer.bricks.snapps.games.go.impl.network.RemotePlayerOnSneer;
+import sneer.bricks.snapps.games.go.impl.sneerSpecifics.RemotePlayerOnSneer;
+import sneer.bricks.snapps.games.go.impl.sneerSpecifics.SneerTimerFactory;
 import basis.lang.ByRef;
 import basis.lang.Closure;
 import basis.lang.Consumer;
@@ -36,7 +36,7 @@ class GoMainImpl implements GoMain {
 
 	private Register<Move> _moveRegister = my(Signals.class).newRegister(null);
 
-	@SuppressWarnings("unused") private final WeakContract _refToAvoidGc;
+	@SuppressWarnings("unused") private WeakContract _refToAvoidGc;
 	@SuppressWarnings("unused") private WeakContract _refToAvoidGc2;
 
 	private Move _remoteMove;
@@ -45,7 +45,10 @@ class GoMainImpl implements GoMain {
 
 	{
 		setupContextMenu();
-		
+		subscribeToReceiveGameMessages();
+	}
+
+	private void subscribeToReceiveGameMessages() {
 		_refToAvoidGc = my(TupleSpace.class).addSubscription(GoMessage.class, new Consumer<GoMessage>() { @Override public void consume(final GoMessage message) {
 			if (message instanceof GoInvitation) { // Why not use only GoMessage with a string field (for handshake and moves exchange)?
 				handleInviation((GoInvitation)message);
@@ -55,7 +58,6 @@ class GoMainImpl implements GoMain {
 				handleMove((GoMove)message);
 			}
 		}});
-
 	}
 
 	private void setupContextMenu() {
@@ -112,12 +114,12 @@ class GoMainImpl implements GoMain {
 		//			my(TimeboxedEventQueue.class).startQueueing(5000); // Fix: Talk to Klaus about Timebox issue
 		my(GuiThread.class).invokeAndWaitForWussies(new Closure(){@Override public void run() {
 			Player remotePlayer = new RemotePlayerOnSneer(_moveRegister);
-			GoFrame goFrame = new GoFrame(stoneColor.value, 0, new SneerTimerFactory());
-			goFrame.setAdversary(remotePlayer);
-			remotePlayer.setAdversary(goFrame);
+			Player localPlayer = new GuiPlayer(stoneColor.value, 0, new SneerTimerFactory()); 
+			localPlayer.setAdversary(remotePlayer);
+			remotePlayer.setAdversary(localPlayer);
 		}});
 	}	
-	
+	 
 	private void handleMove(GoMove message) {
 		if(message.publisher.equals(my(OwnSeal.class).get().currentValue()))
 			return;
