+++ code/src/sneer/bricks/snapps/games/go/impl/Player.java
@@ -1,4 +1,4 @@
-package sneer.bricks.snapps.games.go.impl.network;
+package sneer.bricks.snapps.games.go.impl;
 
 import sneer.bricks.snapps.games.go.impl.logic.Move;
