+++ code/src/sneer/bricks/snapps/games/go/impl/TimerFactory.java
@@ -1,4 +1,4 @@
-package sneer.bricks.snapps.games.go.impl.gui;
+package sneer.bricks.snapps.games.go.impl;
 
 import sneer.bricks.hardware.cpu.lang.contracts.WeakContract;
