+++ code/src/sneer/bricks/snapps/games/go/impl/logic/GoBoard.java
@@ -7,7 +7,6 @@ import java.util.Arrays;
 import java.util.HashSet;
 import java.util.Set;
 
-import sneer.bricks.snapps.games.go.impl.gui.BoardListener;
 
 public class GoBoard {
